Add support for null image backend type
This backend type can be used if we want snf-image to configure the
VM disk without first copying an image to it. This is useful for
cases were the VM disk is already populated before snf-image runs.

+++ snf-image-host/copy-monitor.py
@@ -145,7 +145,7 @@ def main():
         for l in iof.readlines():
             if l.startswith("rchar:"):
                 message['position'] = int(l.split(': ')[1])
-                message['progress'] = float(100) if opts.read_bytes == 0 \
+                message['progress'] = float(0) if opts.read_bytes == 0 \
                     else float("%2.2f" % (
                         message['position'] * 100.0 / message['total']))
                 send_message(out, message)
